Add files via upload
these files are the source code files of a game called Reversi. the code includes an AI so that you could play the game against the computer as well as against another human being. The "test" classes were required to check if each function is working separately. The main functions were implemented in reversi.cpp.